refactor: update architecture for FAISS and sentence-transformers
- Redesign folder structure to a clean `src/` and `scripts/` layout.
- Adapt Flask app paths to new `src/api/` and `src/web/` locations.
- Remove old basic similarity processing logic and models.
- Create placeholder skeleton files for FAISS and SentenceTransformers.
- Update `requirements.txt` with pinned versions for new stack.
- Write a comprehensive `README.md` documenting the new architecture.
