--- app/models/__init__.py
@@ -1,5 +1,0 @@
-from .inference import recommendation_system
-
-__all__ = [
-    'recommendation_system',
-]
