--- app/preprocessing/__init__.py
@@ -1,9 +1,0 @@
-from .text_processing import pipeline
-from .embeddings import get_average_embeddings
-from .data_filtering import get_filtered_df
-
-__all__ = [
-    'pipeline',
-    'get_average_embeddings',
-    'get_filtered_df'
-]
