--- data/__init__.py
@@ -1,7 +1,0 @@
-from .loader import get_dataset
-from .loader import get_pretrained_model
-
-__all__ = [
-    'get_dataset',
-    'get_pretrained_model',
-]
