+++ requirements.txt
@@ -1,6 +1,5 @@
-# Core Libraries
 flask==3.1.0
-numpy==1.26.4
+faiss-cpu==1.9.0
+sentence-transformers==3.3.1
 pandas==2.2.3
-nltk==3.9.1
-gensim==4.3.3
+numpy==1.26.4
