+++ run.py
@@ -1,4 +1,4 @@
-from app.main.app import app
+from src.api.app import app
 
 if __name__ == '__main__':
     app.run(debug=True)
